requirements.txt
django
psycopg[binary]
python-dotenv>=1.0.0
gunicorn
disposable-email-domains
whitenoise==6.7.0
djangorestframework>=3.14.0
drf-spectacular>=0.27.0